+++ Modules/xNetworking/Examples/Resources/xDnsServerAddress/2-xDnsServerAddress_PrimaryAndSecondary.ps1
@@ -1,0 +1,24 @@
+<#
+    .EXAMPLE
+    Configure primary and secondary DNS Server addresses on the Ethernet adapter
+#>
+configuration Example
+{
+    param
+    (
+        [string[]] $NodeName = 'localhost'
+    )
+
+    Import-DscResource -Module xNetworking
+
+    Node $NodeName
+    {
+        xDnsServerAddress DnsServerAddress
+        {
+            Address        = '10.0.0.2','10.0.0.40'
+            InterfaceAlias = 'Ethernet'
+            AddressFamily  = 'IPv4'
+            Validate       = $true
+        }
+    }
+}