+++ Modules/xNetworking/Examples/Resources/xFirewall/2-xFirewall_AddFirewallRule_AllParameters.ps1
@@ -4,7 +4,7 @@
     Note: This configuration sample uses all Firewall rule parameters.
     It is only used to show example usage and should not be created.
 #>
-configuration Example
+Configuration Example
 {
     param
     (