+++ Modules/xNetworking/Examples/Resources/xNetAdapterBinding/1-DisableIPv6.ps1
@@ -2,8 +2,7 @@
     .EXAMPLE
     Disabling IPv6 for the Ethernet adapter
 #>
-
-configuration Example
+Configuration Example
 {
     param
     (