+++ Modules/xNetworking/Examples/Resources/xNetBIOS/1-DisableNetBIOS.ps1
@@ -2,7 +2,7 @@
     .EXAMPLE
     Disable NetBios on Adapter
 #>
-configuration Example
+Configuration Example
 {
     param
     (