+++ Modules/xNetworking/Examples/Resources/xNetConnectionProfile/1-SetNetConnectionProfile.ps1
@@ -2,7 +2,7 @@
     .EXAMPLE
     Sets the Ethernet adapter to Public and IPv4/6 to Internet Connectivity
 #>
-configuration Example
+Configuration Example
 {
     param
     (