+++ Modules/xNetworking/Examples/Resources/xDhcpClient/1-EnableDHCP.ps1
@@ -2,8 +2,7 @@
     .EXAMPLE
     Enabling DHCP Client for the Ethernet Alias
 #>
-
-configuration Example
+Configuration Example
 {
     param
     (