+++ Modules/xNetworking/Examples/Resources/xDnsServerAddress/1-xDnsServerAddress.ps1
@@ -2,7 +2,7 @@
     .EXAMPLE
     Configure DNS Server for the Ethernet adapter
 #>
-configuration Example
+Configuration Example
 {
     param
     (