+++ Modules/xNetworking/Examples/Resources/xFirewall/4-xFirewall_EnableBuiltInFirewallRule.ps1
@@ -7,7 +7,9 @@ Configuration Example
 {
     param
     (
-        [string[]] $NodeName = 'localhost'
+        [Parameter()]
+        [System.String[]]
+        $NodeName = 'localhost'
     )
 
     Import-DSCResource -ModuleName xNetworking