+++ Modules/xNetworking/Examples/Resources/xHostsFile/1-AddEntry.ps1
@@ -6,7 +6,8 @@ Configuration Example
 {
     param
     (
-        [string[]]
+        [Parameter()]
+        [System.String[]]
         $NodeName = 'localhost'
     )