+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Converted to use AppVeyor.psm1 in DSCResource.Tests repository.
+- Converted to use Example and Markdown tests in DSCResource.Tests repository.
+
 ## 3.2.0.0
 
 - Fixed typo in the example's Action property from "Blocked" (which isn't a valid