+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Converted to use AppVeyor.psm1 in DSCResource.Tests repository.
 - Converted to use Example and Markdown tests in DSCResource.Tests repository.
+- Added CodeCov.io support.
 
 ## 3.2.0.0