+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added CodeCov.io support.
 - Added a new example to xDNSServerAddress to clarify setting multiple DNS Servers.
 - Fix examples to correct display in auto documentation generation.
+- Correct example parameters format to meet style guidelines.
 
 ## 3.2.0.0