+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Converted to use AppVeyor.psm1 in DSCResource.Tests repository.
 - Converted to use Example and Markdown tests in DSCResource.Tests repository.
 - Added CodeCov.io support.
+- Added a new example to xDNSServerAddress to clarify setting multiple DNS Servers.
 - Fix examples to correct display in auto documentation generation.
+- Correct example parameters format to meet style guidelines.
 
 ## 3.2.0.0