+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Converted to use AppVeyor.psm1 in DSCResource.Tests repository.
 - Converted to use Example and Markdown tests in DSCResource.Tests repository.
 - Added CodeCov.io support.
+- Added a new example to xDNSServerAddress to clarify setting multiple DNS Servers.
 - Fix examples to correct display in auto documentation generation.
 - BREAKING CHANGE: Migrated xNetworkAdapter module functionality to xNetAdapterName
   resource.
@@ -12,6 +13,7 @@
 - MSFT_xNetAdapterName:
   - Created new resource for renaming network adapters.
   - Added Find-NetAdapter cmdlet to NetworkingDsc.Common.
+- Correct example parameters format to meet style guidelines.
 
 ## 3.2.0.0